chore: Change .gitignore style
Change .gitignore style to "inclusive" style.

In this style, .gitignore declares "which will be included in the repo".
Instead of normal = exclusive style : "which will be excluded from the repo".

Since this project aims to support many platforms,
it may need to ignore countless number of intermediate files / directories for platform X, Y, Z, ...
But with ".gitignore inclusive" style, we can reduce this (future) complexity.
Because we just add "what we actually need".